Holy Lotus Yoga CSS Flexbox Bootstrap 5 Project
This is my second CSS Flexbox project for Codeacademy. I decided to  take the opportunity to learn as  much as possible about Bootstrap 5, so there is a lot going on.But it's a good example of my design and CSS skills. There's some Javascript, but this project does not showcase those skills.